Recompile qub_load/saveTree.mex and miltreeiface executable exclusively using static libraries on both Linux and Windows to reduce potential linking errors at run time. Also used the latest qub-express source code to fix any bugs upstream.
Also remove trailing zero-state dwells that were preventing MIL from
converging. The existing code was incorrect.

+++ qub/qubtree/istream_readline.cpp
@@ -1,5 +1,23 @@
+/* Copyright 1998-2011 Research Foundation State University of New York */
+
+/* This file is part of QuB.                                            */
+
+/* QuB is free software; you can redistribute it and/or modify          */
+/* it under the terms of the GNU General Public License as published by */
+/* the Free Software Foundation, either version 3 of the License, or    */
+/* (at your option) any later version.                                  */
+
+/* QuB is distributed in the hope that it will be useful,               */
+/* but WITHOUT ANY WARRANTY; without even the implied warranty of       */
+/* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the        */
+/* GNU General Public License for more details.                         */
+
+/* You should have received a copy of the GNU General Public License,   */
+/* named LICENSE.txt, in the QuB program directory.  If not, see        */
+/* <http://www.gnu.org/licenses/>.                                      */
+
 #include "istream_readline.h"
-#include <cstring>
+#include <string.h>
 
 // no lame length restrictions here:
 //   keeps doubling buflen until it fits; returns realloc'd buffer
@@ -31,7 +49,7 @@ char *istream_readline( istream& in, char *buf, int &buflen, int &linelen ){
 	if ( in )
 		in.get(newline);
 	
-	linelen = strlen( buf );
+	linelen = (int) strlen( buf );
 	if ( buf[ linelen - 1 ] == '\r' ) {
 		buf[ linelen - 1 ] = 0;
 		linelen--;
